+++ assets/javascripts/application.js
@@ -1,0 +1,6 @@
+//= require javascripts/app/app.js
+//= require javascripts/app/services.js
+//= require javascripts/app/controllers.js
+//= require javascripts/app/filters.js
+//= require javascripts/app/directives.js
+//= require javascripts/main.js