+++ assets/javascripts/vendor.js
@@ -1,0 +1,8 @@
+//= require angular-commons/envs.js
+//= require javascripts/vendor/jquery-1.8.2.js
+//= require javascripts/vendor/bootstrap.js
+//= require javascripts/vendor/angular/angular.js
+//= require javascripts/vendor/angular/angular-cookies.js
+//= require javascripts/vendor/socialite.js
+//= require angular-commons/controllers.js
+//= require angular-commons/services.js