+++ lib/spiral-galaxy.rb
@@ -4,10 +4,6 @@ module Spiral
   end
 end
 
-require 'brochure'
-require 'coffee_script'
-require 'less'
-
 require 'angular-commons-middleware'
 require 'auth-client'
 require 'devcenter-client'
@@ -16,4 +12,5 @@ require "spiral-galaxy/version"
 require "spiral-galaxy/connection"
 require "spiral-galaxy/auth_backend"
 require "spiral-galaxy/canvas_redirect"
+require "spiral-galaxy/assets"
 require "spiral-galaxy/app"