+++ lib/spiral-galaxy/app.rb
@@ -1,5 +1,8 @@
 ENV_KEYS_TO_EXPOSE = ['QS_DEVCENTER_BACKEND_URL', 'QS_PLAYERCENTER_BACKEND_URL', 'QS_CANVAS_APP_URL', 'QS_S3_HOST', 'QS_AUTH_BACKEND_URL']
 
+require 'sprockets'
+require 'rack/cache'
+
 module Spiral::Galaxy
   class App
     def initialize
@@ -16,15 +19,25 @@ module Spiral::Galaxy
           run auth_callback
         end
 
+        if ENV['RACK_ENV'] == 'production'
+          memcached_base_url = "memcached://#{ENV['MEMCACHIER_USERNAME']}:#{ENV['MEMCACHIER_PASSWORD']}@#{ENV['MEMCACHIER_SERVERS']}"
+          use Rack::Cache,
+            :metastore   => "#{memcached_base_url}/meta",
+            :entitystore => "#{memcached_base_url}/body"
+        end
+
         map "/play/" do
           run CanvasRedirect.new
         end
 
-        root = File.expand_path('../../../', __FILE__)
-        brochure = Brochure.app(root)
-        use Angular::Commons::Middleware
+        sprockets_environment = Assets.new
+        index_url = Assets::AssetUrlBuilder.new(sprockets_environment).asset_url('index.html')
 
-        run brochure
+        app = lambda do |env|
+          env['PATH_INFO'] = index_url if env['PATH_INFO'] == '/'
+          sprockets_environment.call(env)
+        end
+        run app
       end
     end